requirements.txt
fastapi
uvicorn
pandas
joblib
sqlalchemy
python-jose[cryptography]
redis
fastapi-limiter
xgboost
python-multipart
pytest
httpx
scikit-learn